Time: 1 ms (76.45%), Space: 42.3 MB (63.60%) - LeetHub

+++ 50-powx-n/50-powx-n.java
@@ -14,7 +14,7 @@ class Solution {
                     }
             }
             if(n<0){
-                    ans=1.0/(double)ans;
+                    ans=1.0/ans;
             }
             return ans;
